More work. Check commit for details.

+++ base/CUSBKeyboard.h
@@ -14,7 +14,7 @@ public:
 	void PressKey(uint8_t key);
 	void PressKey(uint8_t key, uint8_t modifier);
 	void PressKey(uint8_t key, uint8_t modifier, uint8_t repeatCount);
-	void PressKey(uint8_t key, uint8_t modifier, uint8_t repeatCount, uint8_t modifierDelay);
+	void PressKey(uint8_t key, uint8_t modifier, uint8_t repeatCount, uint16_t modifierDelay);
 	void TypeString(const char* string);
 	void TypeChar(char c);
